requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt
uuid